part_disappear: bridge + DisappearPhase/Timer system (tests: boundaries, order, no double fire); tradeoffs entry (a)
Milestone V3-A commit 4 (tasks T022–T027). part_disappear/system.rs holds DisappearPhase
(EMIT_DELAY = 0.2, LIFETIME_FACTOR = 2.0, start(), step()), Lifetime, Transition and the
`advance` gameplay system (Res<FrameDelta>, Commands); part_disappear.rs is a bridge (ready
keeps MiniBlasts.emitting = true, registers Handles::Puff with the root upcast to
CpuParticles3D and Initial::Puff { lifetime: get_lifetime() }; exit_tree unregisters; no
async block, no process); sync_out_puff in the FRAME SyncOut before sync_out_remove; advance
in build_frame's Gameplay; apply_register inserts DisappearPhase::start() + Lifetime.
docs/v3-tradeoffs.md row (a).

Gate: cargo build / clippy (0 warnings) / cargo test = 156 passed. Headless: import OK;
main/main.tscn x2 and level/level.tscn clean vs baseline.

Harness case (b), v2 worktree (e2932b4) vs v3, --fixed-fps 60 --quit-after 400, split
XDG_DATA_HOME, puff instanced from a SceneTreeTimer(0.5).timeout callback: 402 lines each.
`diff` of the full logs:
  227c227
  < RAW F225 puff_tree_exited
  ---
  > RAW F224 puff_tree_exited
`diff` of the '^RAW' subsets: the same single line (v2 F225, v3 F224). Every observer line is
identical on both trees: puff_in_tree false F0-F30; in tree, puff_emitting=false,
mini_emitting=true F31-F43; puff_emitting=true F44-F48 (mini still true); mini false
F49-F133; puff_emitting=false F134-F224; out of tree F225-F399. So the measured timing
difference for the spec's table (Session 4) is: puff free RAW stamp v2 = F225, v3 = F224,
delta -1 frame on v3, observer-invisible — exactly research R5's prediction (v2's
godot::task future resumes via call_deferred in the next iteration's physics phase).

Deviation (signature mismatch, reported): gdext's CpuParticles3D::get_lifetime() returns f64
(generated bindings cpu_particles_3d.rs:308), not the f32 data-model.md assumed; v2's
`lifetime * 2.0` was therefore an f64 product. Lifetime(f64), Initial::Puff { lifetime: f64 }
and LIFETIME_FACTOR: f64 reproduce v2 exactly; the "f32 multiply then widen" rationale in
data-model.md/tasks T022 is wrong and needs a one-line correction (Session 4 docs pass).
`mod system;` is `pub(crate) mod system;` (reachable from ecs/setup.rs), as in commit 3.

+++ oxide_godot_core/oxide_godot_lib/src/ecs/event.rs
@@ -32,8 +32,10 @@ pub enum Initial {
     /// → `DoorState::Closed`
     Door,
     /// → `DisappearPhase::start()` + `Lifetime(lifetime)`; `lifetime` is the node's
-    /// `CPUParticles3D.lifetime` read once in `ready` (v2 `part_disappear.rs:35`, an `f32`).
-    Puff { lifetime: f32 },
+    /// `CPUParticles3D.lifetime` read once in `ready` (v2 `part_disappear.rs:35`; the gdext
+    /// binding `CpuParticles3D::get_lifetime` returns `f64`, so v2's `lifetime * 2.0` was an
+    /// `f64` product — implement-time correction of data-model.md's `f32`).
+    Puff { lifetime: f64 },
     /// → `BlastTag`
     Blast,
 }
